+++ src/demo/org/jdbx/demo/UpdateDemo.java
@@ -27,7 +27,7 @@ public class UpdateDemo
 	public static void stmt(StaticStmt stmt) throws JdbxException
 	{
 		String sql1 	= "INSERT INTO X VALUES('a', 'b')";
-		int count  		= stmt.createUpdate(sql1).returnAutoKeyCols().run().count(); // todo receive autokeys
+		long count  	= stmt.createUpdate(sql1).returnAutoKeyCols().run().count(); // todo receive autokeys
 
 		String sql2		= "SELECT amount FROM Orders WHERE id = 1";
 		double amount 	= stmt.query(sql2).row().col().getDouble();
